Detect pyenv python installations, and ignore (presumed) python installations that don't have semver formatting.
Before, an executable with a name like `python-config` would be detected as python, and it would be assumed that `python-config --version` would return a version in the form of `X.XX.X`. However, this is not always the case with custom commands.

Additionally, pyenv installations were not detected, which is common on some systems that only have one python version in their repo.

Signed-off-by: carnivorecookies <carnivorecookies@proton.me>